:wrench: [GPP-Woo/GPP-zoeken#81] Copy over supported file types
Copied the list of supported file types from Felix' comment at
https://github.com/GPP-Woo/GPP-zoeken/pull/89\#issuecomment-2893506130
and added an additional configuration key so that GPP-Zoeken can
filter out only the relevant file types for indexing.

The purpose of this file is now also documented in the README, as
JSON files themselves don't allow for comments.